[TAO-1796][Refactor] video_clip: move the decode backends into video_decode
video_text_loader.py had grown to 1449 lines covering metadata parsing, video
decoding, tensor stacking, the Dataset and the DataLoader factory. The PyAV
switch added eight more private functions to it, and decoding is the concern
least related to "video-text dataset support".

Move the decode chain to dataloader/video_decode.py with load_video_frames as
its only public symbol: the PyAV backend and its helpers, the ffmpeg CLI
backend and its ffprobe helper, the OpenCV backend, the shared
_linspace_indices/_clip_frame_range range helpers, and the backend logger.

Pure move -- no function body is edited. video_text_loader re-exports
load_video_frames, so scripts/inference.py and the test import block are
unchanged, and monkeypatching video_text_loader.load_video_frames still works
because "from X import Y" binds the name locally. The only orphan the move left
behind, an unused PIL.Image import, is dropped.

video_text_loader.py 1449 -> 996 lines; video_decode.py 482.

Verified: flake8 (CI flags) clean, pylint 10.00, pydocstyle clean,
86 passed / 1 skipped in tests/multimodal_unit_test/video_clip/, and both
import paths resolve to the same function object.
